backend/requirements.txt
flask
flask-cors
flask-mysqldb
mysql-connector-python
python-dotenv
bcrypt
pytz
